Rename transactions extension to narayana-jta
