+++ bom/pom.xml
@@ -146,7 +146,7 @@
 
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-transactions-deployment</artifactId>
+            <artifactId>shamrock-narayana-jta-deployment</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
          </dependency>
