+++ build-parent/pom.xml
@@ -403,12 +403,12 @@
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-transactions-deployment</artifactId>
+            <artifactId>shamrock-narayana-jta-deployment</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-transactions-runtime</artifactId>
+            <artifactId>shamrock-narayana-jta-runtime</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
