+++ extensions/narayana-jta/runtime/src/main/java/org/jboss/shamrock/narayana/jta/runtime/NarayanaJtaConfiguration.java
@@ -1,0 +1,17 @@
+package org.jboss.shamrock.narayana.jta.runtime;
+
+import org.jboss.shamrock.runtime.annotations.ConfigItem;
+import org.jboss.shamrock.runtime.annotations.ConfigPhase;
+import org.jboss.shamrock.runtime.annotations.ConfigRoot;
+
+/**
+ *
+ */
+@ConfigRoot(phase = ConfigPhase.RUN_TIME_STATIC)
+public final class NarayanaJtaConfiguration {
+    /**
+     * The node name used by the transaction manager
+     */
+    @ConfigItem(defaultValue = "shamrock")
+    public String nodeName;
+}
