+++ extensions/narayana-jta/runtime/src/main/java/org/jboss/shamrock/narayana/jta/runtime/NarayanaJtaProducers.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2018 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.shamrock.narayana.jta.runtime;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.Dependent;
+import javax.enterprise.inject.Produces;
+import javax.transaction.TransactionSynchronizationRegistry;
+
+import org.jboss.tm.JBossXATerminator;
+import org.jboss.tm.XAResourceRecoveryRegistry;
+import org.jboss.tm.usertx.UserTransactionRegistry;
+
+import com.arjuna.ats.internal.jbossatx.jta.jca.XATerminator;
+import com.arjuna.ats.internal.jta.transaction.arjunacore.TransactionSynchronizationRegistryImple;
+import com.arjuna.ats.jbossatx.jta.RecoveryManagerService;
+import com.arjuna.ats.jta.TransactionManager;
+import com.arjuna.ats.jta.UserTransaction;
+
+@Dependent
+public class NarayanaJtaProducers {
+
+    @Produces
+    @ApplicationScoped
+    public UserTransactionRegistry userTransactionRegistry() {
+        return new UserTransactionRegistry();
+    }
+
+    @Produces
+    @ApplicationScoped
+    public javax.transaction.UserTransaction userTransaction() {
+        return UserTransaction.userTransaction();
+    }
+
+    @Produces
+    @ApplicationScoped
+    public XAResourceRecoveryRegistry xaResourceRecoveryRegistry() {
+        return new RecoveryManagerService();
+
+    }
+
+    @Produces
+    @ApplicationScoped
+    public TransactionSynchronizationRegistry transactionSynchronizationRegistry() {
+        return new TransactionSynchronizationRegistryImple();
+    }
+
+    @Produces
+    @ApplicationScoped
+    public javax.transaction.TransactionManager transactionManager() {
+        return TransactionManager.transactionManager();
+    }
+
+    @Produces
+    @ApplicationScoped
+    public JBossXATerminator xaTerminator() {
+        return new XATerminator();
+    }
+}
