+++ extensions/narayana-jta/runtime/src/main/java/org/jboss/shamrock/narayana/jta/runtime/graal/PropertiesFactorySubstitution.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2018 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.shamrock.narayana.jta.runtime.graal;
+
+import java.util.Properties;
+
+import org.jboss.shamrock.narayana.jta.runtime.NarayanaJtaTemplate;
+
+import com.arjuna.common.util.propertyservice.PropertiesFactory;
+import com.oracle.svm.core.annotate.Substitute;
+import com.oracle.svm.core.annotate.TargetClass;
+
+@TargetClass(PropertiesFactory.class)
+final class PropertiesFactorySubstitution {
+
+    @Substitute
+    public static Properties getDefaultProperties() {
+        return NarayanaJtaTemplate.getDefaultProperties();
+    }
+
+}
